+++ CHANGELOG.md
@@ -46,6 +46,7 @@
   * Fixed floating building in Town03
   * Fixed vehicles missing the route if autopilot enabled too late
   * Enhanced stop triggers options
+  * Fixed division by zero in is_within_distance_ahead()
 
 ## CARLA 0.9.4